Changed some models for WebSocket requests and responses Fully configured web socket connection

+++ src/main/java/me/gnevilkoko/threads/StreamThread/models/DonationAlertsAccesTokenConfig.java
@@ -72,8 +72,8 @@ public class DonationAlertsAccesTokenConfig extends YamlModel {
     public String getSocketConnectionToken(){
         return (String) getData().get("socket_connection_token");
     }
-    public long getUserID(){
-        return (long) getData().get("user_id");
+    public int getUserID(){
+        return (int) getData().get("user_id");
     }
 
     public enum Names {
